PCI-1595 Refactor Dependabot docker to allow other package managers
