+++ ci/tasks/test-dependabot-task.yml
@@ -27,9 +27,11 @@ run:
 
       bundle config build.nokogiri --use-system-libraries
       bundle config set path '../../requirements'
-      cd dependabot-core.git/docker
+      cd dependabot-core.git/pix4d-dependabot
       bundle install
-      bundle exec rspec ./spec/
       bundle exec rubocop -c ../.rubocop.yml ../
+      cd ..
+      cd docker && bundle exec rspec . && cd -
+      cd pix4d-dependabot && bundle exec rspec . && cd -
       export COMMIT_HASH=$(cat ../.git/ref)
       bundle exec ruby ../ci/tasks/pr_auto_merge.rb
