+++ pix4d-dependabot/.gitignore
@@ -1,0 +1,1 @@
+run_script.sh
