+++ pix4d-dependabot/.rubocop.yml
@@ -1,0 +1,1 @@
+inherit_from: ../.rubocop.yml
