+++ pix4d-dependabot/Gemfile
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+source "https://rubygems.org"
+
+gem "dependabot-common", path: "../common"
+gem "dependabot-docker", path: "../docker"
+
+gemspec
