+++ pix4d-dependabot/lib/helpers/auto_merge.rb
